Backend/requirement.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.40.0
pinecone>=5.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0
pypdf>=4.0.0
python-dotenv>=1.0.0
